+++ Vector.h
@@ -3,18 +3,18 @@
 
 struct Vec2D {
 public:
-  Vec2D(const double x_ = 0, const double y_ = 0) : x(x_), y(y_) {}
-  double x;
-  double y;
+  Vec2D(const float x_ = 0, const float y_ = 0) : x(x_), y(y_) {}
+  float x;
+  float y;
 };
 
 struct Vec3D {
 public:
-  Vec3D(const double x_ = 0, const double y_ = 0, const double z_ = 0)
+  Vec3D(const float x_ = 0, const float y_ = 0, const float z_ = 0)
       : x(x_), y(y_), z(z_) {}
-  double x;
-  double y;
-  double z;
+  float x;
+  float y;
+  float z;
 };
 
 #endif // __GALAXY_RENDERER_VECTOR_H__
